Moved backbone-abort-sync as a util
